remove system_rs crate
