+++ Cargo.toml
@@ -17,7 +17,6 @@ thiserror = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 serde_with = { version = "3.2", features = ["json"] }
-system_rs = { path = "./crates/system_rs" }
 
 [package]
 edition = "2021"
@@ -61,7 +60,6 @@ tracing-subscriber-wasm.workspace = true
 uuid = { version = "1.3.0", features = ["serde", "v4", "js"] }
 wasm-bindgen.workspace = true
 nix_rs.workspace = true
-system_rs.workspace = true
 nix_health.workspace = true
 leptos_extra.workspace = true
