+++ crates/nix_health/src/check/min_nix_version.rs
@@ -11,7 +11,7 @@ use crate::{
 pub struct MinNixVersion(pub NixVersion);
 
 impl Check for MinNixVersion {
-    fn check(nix_info: &info::NixInfo, _sys_info: &system_rs::info::SysInfo) -> Self {
+    fn check(nix_info: &info::NixInfo, _sys_info: &info::SysInfo) -> Self {
         MinNixVersion(nix_info.nix_version.clone())
     }
     fn name(&self) -> &'static str {
