+++ crates/nix_health/src/traits.rs
@@ -1,6 +1,5 @@
 use crate::report::{Report, WithDetails};
 use nix_rs::info;
-use system_rs;
 
 /// Types that implement health check with reports
 pub trait Check {
@@ -8,7 +7,7 @@ pub trait Check {
     type Report = Report<WithDetails>;
 
     /// Run and create the health check
-    fn check(nix_info: &info::NixInfo, sys_info: &system_rs::info::SysInfo) -> Self
+    fn check(nix_info: &info::NixInfo, sys_info: &info::SysInfo) -> Self
     where
         Self: Sized;
