+++ crates/nix_rs/Cargo.toml
@@ -11,6 +11,7 @@ crate-type = ["cdylib", "rlib"]
 [dependencies]
 cfg-if.workspace = true
 regex = "1.9.3"
+os_info = "3.7.0"
 thiserror.workspace = true
 serde.workspace = true
 serde_json.workspace = true
