+++ crates/nix_rs/src/info.rs
@@ -1,5 +1,8 @@
-//! Information about the user's Nix installation
+//! Information about the user's Nix installation and system
+use os_info;
 use serde::{Deserialize, Serialize};
+#[cfg(feature = "all")]
+use std::env;
 
 use crate::{config::NixConfig, version::NixVersion};
 
@@ -23,5 +26,23 @@ impl NixInfo {
             nix_version,
             nix_config,
         })
+    }
+}
+
+/// Information about the user's system
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct SysInfo {
+    /// value of $USER
+    pub current_user: String,
+    pub os: os_info::Type,
+}
+
+impl SysInfo {
+    /// Determine [SysInfo] on the user's system
+    #[cfg(feature = "all")]
+    pub async fn get_info() -> Result<SysInfo, env::VarError> {
+        let current_user = env::var("USER")?;
+        let os = os_info::get().os_type();
+        Ok(SysInfo { current_user, os })
     }
 }
