--- crates/system_rs/src/lib.rs
@@ -1,4 +1,0 @@
-#![feature(associated_type_defaults)]
-//! Rust crate to get the user system info
-//!
-pub mod info;
