+++ src/app/health.rs
@@ -16,8 +16,6 @@ use nix_health::{
     NixHealth,
 };
 use nix_rs::version::NixVersion;
-#[cfg(feature = "ssr")]
-use system_rs;
 use tracing::instrument;
 
 use crate::{app::info::ConfigValListView, widget::*};
@@ -169,7 +167,7 @@ pub async fn get_nix_health(_unit: ()) -> Result<nix_health::NixHealth, ServerFn
     use nix_health::{traits::Check, NixHealth};
     use nix_rs::info;
     let nix_info = info::NixInfo::from_nix(&nix_rs::command::NixCmd::default()).await?;
-    let sys_info = system_rs::info::SysInfo::get_info().await?;
+    let sys_info = info::SysInfo::get_info().await?;
     let health = NixHealth::check(&nix_info, &sys_info);
     Ok(health)
 }
